[btrie] BUGFIX Right redistribution
Right redistribution on insert was broken: When the key being added was
on the boundary (and would become the first element in the right hand
node) the partitioning key was being discarded because an error in
book-keeping. The `keep` variable should always have the value of
`node->count - delta`, but when delta was decremented it was not
updated. To remedy this, `keep` was removed altogether and all instances
of its usage were replaced with `node->count - delta`.